add etckeeper {pre,post}run_commands to generated puppet.conf
This allows automatic integration with the etckeeper system, if it is present
and configured.  AFAIK this is the only way available to track /etc changes
pre/post a puppet apply run; but even if not it's very common and useful :)

The etckeeper scripts are supplied (and enabled by default) by the ubuntu
puppet-common package.